# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Gollege",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/enem/files/{year}/{id}": {
            "get": {
                "description": "Get file from an specific year and id",
                "produces": [
                    "*/*"
                ],
                "tags": [
                    "ENEM"
                ],
                "summary": "Get file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Year",
                        "name": "year",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/enem/{year}/{lang}": {
            "get": {
                "description": "List question from an specific year and language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ENEM"
                ],
                "summary": "List questions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Year",
                        "name": "year",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}